requirements.txt
argparse
capstone
impacket
keystone-engine
psutil
pycryptodomex
pyrdp
pyfiglet
termcolor